ci: rename pathlib_ext to common_ext
as functions are more than PathLib extension and new ones
will be added.
